The recommendedModels now returns arrays of model names.
They are ordered from most preferred to least proferred. If available, the first is the default. After that, the supported are arranged from smallest to largest. Generally, taking the first in the list is a good choice.

When language constraints are applied, models that poorly support the language in question are not in the recommended list.